fix(analytics deps): Fixes
 - Refactored analytics stories to storybook project
