+++ packages/analytics/src/components/SingleValueContainer/index.tsx
@@ -5,15 +5,22 @@ import SingleValueItem from "./components/SingleValueItem/SingleValueItem";
 import styles from "./styles/SingleValueContainer.module.css";
 import {SingleValue, SingleValueContainerProps} from "./types/props";
 
-export function SingleValueContainer({ title, singleValueItems, animationDuration, animationDelay }: SingleValueContainerProps): React.ReactElement {
-  return (
-    <div className="w-100 h-100">
-      <CssReset />
-      <Suspense fallback={<div>{i18n.t("Loading ...")}</div>}>
-        <div>
-          <span className={`${styles["font-x-large"]} ${styles["font-bold"]}`}>{title}</span>
-          <div className={styles["single-value-list"]}>
-            {singleValueItems.map((singleValueItem: SingleValue) => (
+export * from "./types/props";
+
+export function SingleValueContainer({
+                                         title,
+                                         singleValueItems,
+                                         animationDuration,
+                                         animationDelay
+                                     }: SingleValueContainerProps): React.ReactElement {
+    return (
+        <div className="w-100 h-100">
+            <CssReset/>
+            <Suspense fallback={<div>{i18n.t("Loading ...")}</div>}>
+                <div>
+                    <span className={`${styles["font-x-large"]} ${styles["font-bold"]}`}>{title}</span>
+                    <div className={styles["single-value-list"]}>
+                        {singleValueItems.map((singleValueItem: SingleValue) => (
               <SingleValueItem
                 key={`${singleValueItem.label}-${singleValueItem.value}`}
                 {...singleValueItem}
