+++ packages/analytics/src/index.ts
@@ -3,3 +3,4 @@ export * from "./components/CircularProgressDashboard";
 export * from "./components/ChartAnalytics";
 export * from "./components/SingleValueContainer";
 export * from "./components/CustomPivotTable"
+export * from "./components/Visualization"
