+++ packages/storybook/src/analytics/map/Map.stories.tsx
@@ -1,7 +1,6 @@
 import type {Story} from "@storybook/react";
 import React from "react";
-import {MapProps} from "./interfaces";
-import {Map} from "./index";
+import {Map, MapProps} from "@hisptz/dhis2-analytics";
 
 const Template: Story<MapProps> = (args) => <Map {...args} />;
